Extract OverlayMessage to own component
